@@ -2,10 +2,11 @@
 
 buildscript {
     repositories {
+        google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.1.3'
+        classpath 'com.android.tools.build:gradle:3.5.2'
         classpath 'com.github.dcendents:android-maven-gradle-plugin:1.4.1'
 
         // NOTE: Do not place your application dependencies here; they belong
@@ -15,6 +16,7 @@ buildscript {
 
 allprojects {
     repositories {
+        google()
         jcenter()
     }
 }
@@ -25,16 +27,16 @@ task clean(type: Delete) {
 
 ext {
     minSdkVersion = 15
-    compileSdkVersion = 24
-    targetSdkVersion = 24
-    buildToolsVersion = "23.0.2"
+    compileSdkVersion = 29
+    targetSdkVersion = 29
+    buildToolsVersion = "29.0.0"
 
     // android
-    appcompat = 'com.android.support:appcompat-v7:24.1.1'
+    appcompat = 'androidx.appcompat:appcompat:1.1.0'
     //supportv4 = 'com.android.support:support-v4:24.1.1'
-    //supportannotations = 'com.android.support:support-annotations:24.1.1'
+    annotation = "androidx.annotation:annotation:1.1.0"
     //recyclerview = 'com.android.support:recyclerview-v7:24.1.1'
-    design = 'com.android.support:design:24.1.1'
+    design = 'com.google.android.material:material:1.0.0'
     //percent = 'com.android.support:percent:24.1.1'
 
 }
